ircmsg: UTF8-aware truncation during parsing
A message that exceeds the length limit is a protocol violation, so handling
is implementation-defined and it's not really a correctness issue for us to
truncate it additionally.

Also move TruncateUTF8Safe into ircmsg (providing an alias in ircutils for
API compatibility).
